Rename the manager to pubsub and add simple test
